Imported communication code to InstanceManager
