+++ Sim_Harness_GUI/MainWindow.cs
@@ -2,22 +2,12 @@
 using Gtk;
 using Sim_Harness_GUI;
 using System.IO;
-using System.Messaging; //message queues
-using System.Diagnostics; //processes
+
 
 
 public partial class MainWindow: Gtk.Window
 {
 	protected InstanceManager _instances;
-	ProcessStartInfo appGenerator_info = new ProcessStartInfo();
-	ProcessStartInfo houseGenerator_info = new ProcessStartInfo();
-	Process appGenerator, houseGenerator;
-	MessageQueue appQueue_r, appQueue_w, houseQueue_r, houseQueue_w;
-	string appQueueName_r = @".\private$\appQueueRead";
-	string appQueueName_w = @".\private$\appQueueWrite";
-	string houseQueueName_r = @".\private$\houseQueueRead";
-	string houseQueueName_w = @".\private$\houseQueueWrite";
-	//NOTE: names are from the parent's (this program's) perspective
 
 	public MainWindow() : base(Gtk.WindowType.Toplevel)
 	{
@@ -30,45 +20,6 @@ public partial class MainWindow: Gtk.Window
 		Console.WriteLine("Deleted");
 		Application.Quit();
 		a.RetVal = true;
-	}
-
-	protected void OnStartTestButtonClicked(object sender, EventArgs e)
-	{
-		String jsonStartString = buildStartString();
-		currentTestTextview.Buffer.Text = jsonStartString;
-
-		//open the message queues
-		openMessageQueue(ref appQueue_r, ref appQueueName_r);
-		openMessageQueue(ref appQueue_w, ref appQueueName_w);
-		openMessageQueue(ref houseQueue_r, ref houseQueueName_r);
-		openMessageQueue(ref houseQueue_w, ref houseQueueName_w);
-
-		//set process settings
-		appGenerator_info.FileName = appSimLocationEntry.Text;
-		houseGenerator_info.FileName = houseSimLocationEntry.Text;
-		appGenerator_info.Arguments = string.Concat(appQueueName_r, " ", appQueueName_w);
-		houseGenerator_info.Arguments = string.Concat(houseQueueName_r, " ", houseQueueName_w);
-
-		//start the processes
-		currentTestTextview.Buffer.Text += "App: ";
-		currentTestTextview.Buffer.Text += startProcess(ref appGenerator, ref appGenerator_info);
-		currentTestTextview.Buffer.Text += "House: ";
-		currentTestTextview.Buffer.Text += startProcess(ref houseGenerator, ref houseGenerator_info);
-
-		//send the first JSON string through the queue
-		sendMessage(ref appQueue_w, ref jsonStartString);
-		sendMessage(ref houseQueue_w, ref jsonStartString);
-
-		//receive confirmation messages
-		currentTestTextview.Buffer.Text += "App: ";
-		currentTestTextview.Buffer.Text += receiveMessage(ref appQueue_r);
-		currentTestTextview.Buffer.Text += "\n";
-		currentTestTextview.Buffer.Text += "House: ";
-		currentTestTextview.Buffer.Text += receiveMessage(ref houseQueue_r);
-		currentTestTextview.Buffer.Text += "\n";
-
-		startTestButton.Sensitive = false;
-		endTestButton.Sensitive = true;
 	}
 
 	protected void OnLoadScenarioButton(object sender, EventArgs e)
@@ -194,69 +145,27 @@ public partial class MainWindow: Gtk.Window
 		return jsonString;
 	}
 
+	protected void OnStartTestButtonClicked(object sender, EventArgs e)
+	{
+		_instances = new InstanceManager();
+		String jsonStartString = buildStartString();
+
+		currentTestTextview.Buffer.Text += "Attempting to open the Generator processes..\n";
+		currentTestTextview.Buffer.Text += _instances.startGeneratorProcesses(appSimLocationEntry.Text, houseSimLocationEntry.Text);
+		currentTestTextview.Buffer.Text += "Sending the JSON string to the Generator processes...\n";
+		currentTestTextview.Buffer.Text += _instances.sendJSON(jsonStartString);
+
+		startTestButton.Sensitive = false;
+		endTestButton.Sensitive = true;
+	}
+
 	protected void OnEndTestButtonClicked (object sender, EventArgs e)
 	{
-		currentTestTextview.Buffer.Text += ("App: ");
-		currentTestTextview.Buffer.Text += (killProcess(ref appGenerator));
-		currentTestTextview.Buffer.Text += ("House: ");
-		currentTestTextview.Buffer.Text += (killProcess(ref houseGenerator));
+		currentTestTextview.Buffer.Text += _instances.killGeneratorProcesses();
 		endTestButton.Sensitive = false;
 		startTestButton.Sensitive = true;
 	}
 
 
-	//  helper functions //
-
-	private string startProcess(ref Process p, ref ProcessStartInfo ps) {
-		string output = "";
-		try {
-			p = Process.Start(ps);
-			output = string.Concat(output,"Process started successfully\n");
-		} catch (Exception ex) {
-			output = string.Concat(output, "Unable to start process from location: ");
-			output = string.Concat(output, ps.FileName);
-			output = string.Concat(output, "\n");
-		}
-		return output;
-	}
-
-	private string killProcess(ref Process p){
-		string output = "";
-		try {
-			p.Kill();
-			output = string.Concat(output,"Process killed successfully\n");
-		} catch (InvalidOperationException ex) {
-			output = string.Concat(output, "InvalidOperationException thrown - the process has already exited\n");
-		} catch (Exception ex) {
-			output = string.Concat(output, "Exception thrown while trying to kill the process\n");
-			output = string.Concat(output, ex.Message);
-			output = string.Concat(output, "\n");
-		}
-		return output;
-	}
-
-	private void openMessageQueue(ref MessageQueue mQueue, ref string mQueue_name){
-		if (!MessageQueue.Exists(mQueue_name))
-			mQueue = MessageQueue.Create(mQueue_name); //create the queue if it doesn't exist
-		else {
-			mQueue = new MessageQueue(mQueue_name); //connect to the existing queue
-			mQueue.Purge(); //delete any messages that might be in the existing queue
-		}
-	}
-
-	private void sendMessage(ref MessageQueue mQueue, ref string message) {
-		System.Messaging.Message messagetosend = new System.Messaging.Message();
-		messagetosend.Body = message;
-		mQueue.Send(messagetosend);
-	}
-
-	private string receiveMessage(ref MessageQueue mQueue) {
-		string output = "";
-		System.Messaging.Message received_message = new System.Messaging.Message();
-		received_message = mQueue.Receive(); //blocking
-		received_message.Formatter = new XmlMessageFormatter(new String[] { "System.String,mscorlib" });
-		output = (string)received_message.Body;
-		return output;
-	}
 }
