+++ Sim_Harness_GUI/gtk-gui/gui.stetic
@@ -216,7 +216,7 @@
                       </widget>
                       <packing>
                         <property name="Position">1</property>
-                        <property name="AutoSize">True</property>
+                        <property name="AutoSize">False</property>
                       </packing>
                     </child>
                   </widget>
